Replace prepare with prepublishOnly
The "prepare" script is run not only when the package is published, but
also when it is *installed*. This means that it'll try to build the
package on the machine of whoever installs it, which sometimes fails.
"prepublishOnly" scripts are only run when publishing the package, which
is what you probably want for running build scripts.